requirements.txt
streamlit
paddleocr
opencv-python
pillow
pandas
streamlit-drawable-canvas
